Change themes completed, creating providers.tsx, creating and separating components ButtonTheme and NavItem to Header component and edit globals styles
