+++ styles/globals.css
@@ -3,8 +3,8 @@
 @tailwind utilities;
 
 ::selection {
-  background-color: #4600de;
-  color: #000000;
+  background-color: #6a00ff;
+  color: #ffffff;
 }
 
 html {
